+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opModes/autos/ObjectDetectionTeamMadeObjects.java
@@ -77,10 +77,10 @@ public class ObjectDetectionTeamMadeObjects extends LinearOpMode {
             NavFunctions.updateDt(nav);
             NavFunctions.updateHeading(nav);
 
-            float PIDOut = PIDFunctions.updatePIDAngular(
+            double PIDOut = PIDFunctions.updatePIDAngular(
                     drivePID,
                     nav.heading,
-                    (float)nav.dt);
+                    nav.dt);
