+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/V2f.java
@@ -1,11 +1,11 @@
 package org.firstinspires.ftc.teamcode.util;
 
 public class V2f {
-    public float x = 0;
-    public float y = 0;
+    public double x = 0;
+    public double y = 0;
 
     public V2f() { }
-    public V2f(float _x, float _y) {
+    public V2f(double _x, double _y) {
         this.x = _x;
         this.y = _y;
     }
@@ -16,31 +16,31 @@ public class V2f {
     }
 
 
-    public V2f rotated(float deg){
-        float degInRad = (float)Math.toRadians(deg);
-        float sin = (float)Math.sin(degInRad);
-        float cos = (float)Math.cos(degInRad);
+    public V2f rotated(double deg){
+        double degInRad = Math.toRadians(deg);
+        double sin = Math.sin(degInRad);
+        double cos = Math.cos(degInRad);
 
         return new V2f(
                 this.x * cos - this.y * sin,
                 this.y * cos + this.x * sin);
     }
 
-    public float getAngleDeg(){
-        return (float)Math.toDegrees(Math.atan2(this.y, this.x));
+    public double getAngleDeg(){
+        return Math.toDegrees(Math.atan2(this.y, this.x));
     }
 
 
     public V2f add(V2f s) {
         return new V2f(this.x + s.x, this.y + s.y); }
 
-    public V2f divide(float s) {
+    public V2f divide(double s) {
         return new V2f(this.x / s, this.y / s); }
 
-    public V2f multiply(float s) {
+    public V2f multiply(double s) {
         return new V2f(this.x * s, this.y * s); }
-    public float length(){
-        return (float)Math.sqrt(this.x*this.x + this.y*this.y);
+    public double length(){
+        return Math.sqrt(this.x*this.x + this.y*this.y);
     }
 
     public V2f normalized() {
